+++ lib/widget/stock_detail.dart
@@ -567,7 +567,7 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
       child: ListTile(
         title: Text("${widget.item.partName}"),
         subtitle: Text("${widget.item.partDescription}"),
-        leading: InvenTreeAPI().getImage(widget.item.partImage),
+        leading: InvenTreeAPI().getThumbnail(widget.item.partImage),
         trailing: Text(
           api.StockStatus.label(widget.item.status),
           style: TextStyle(
@@ -595,7 +595,8 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
    * Construct a list of detail elements about this StockItem.
    * The number of elements may vary depending on the StockItem details
    */
-  List<Widget> detailTiles() {
+  @override
+  List<Widget> getTiles(BuildContext context) {
     List<Widget> tiles = [];
 
     // Image / name / description
@@ -667,11 +668,7 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
           title: Text(L10().supplierPart),
           subtitle: Text(widget.item.supplierSKU),
           leading: FaIcon(FontAwesomeIcons.building, color: COLOR_ACTION),
-          trailing: InvenTreeAPI().getImage(
-            widget.item.supplierImage,
-            width: 40,
-            height: 40,
-          ),
+          trailing: InvenTreeAPI().getThumbnail(widget.item.supplierImage),
           onTap: () async {
             showLoadingOverlay(context);
             var sp = await InvenTreeSupplierPart().get(
@@ -845,13 +842,4 @@ class _StockItemDisplayState extends RefreshableState<StockDetailWidget> {
     return tiles;
   }
 
-  @override
-  Widget getBody(BuildContext context) {
-    return ListView(
-      children: ListTile.divideTiles(
-          context: context,
-          tiles: detailTiles()
-      ).toList()
-    );
-  }
 }