+++ lib/widget/supplier_part_detail.dart
@@ -102,7 +102,8 @@ class _SupplierPartDisplayState extends RefreshableState<SupplierPartDetailWidge
   /*
    * Build a set of tiles to display for this SupplierPart
    */
-  List<Widget> detailTiles(BuildContext context) {
+  @override
+  List<Widget> getTiles(BuildContext context) {
     List<Widget> tiles = [];
 
     if (loading) {
@@ -116,11 +117,7 @@ class _SupplierPartDisplayState extends RefreshableState<SupplierPartDetailWidge
           title: Text(L10().internalPart),
           subtitle: Text(widget.supplierPart.partName),
           leading: FaIcon(FontAwesomeIcons.shapes, color: COLOR_ACTION),
-          trailing: InvenTreeAPI().getImage(
-            widget.supplierPart.partImage,
-            width: 40,
-            height: 40,
-          ),
+          trailing: InvenTreeAPI().getThumbnail(widget.supplierPart.partImage),
           onTap: () async {
             showLoadingOverlay(context);
             final part = await InvenTreePart().get(widget.supplierPart.partId);
@@ -140,11 +137,7 @@ class _SupplierPartDisplayState extends RefreshableState<SupplierPartDetailWidge
         title: Text(L10().supplier),
         subtitle: Text(widget.supplierPart.supplierName),
         leading: FaIcon(FontAwesomeIcons.building, color: COLOR_ACTION),
-        trailing: InvenTreeAPI().getImage(
-          widget.supplierPart.supplierImage,
-          width: 40,
-          height: 40,
-        ),
+        trailing: InvenTreeAPI().getThumbnail(widget.supplierPart.supplierImage),
         onTap: () async {
           showLoadingOverlay(context);
           var supplier = await InvenTreeCompany().get(widget.supplierPart.supplierId);
@@ -175,11 +168,7 @@ class _SupplierPartDisplayState extends RefreshableState<SupplierPartDetailWidge
           title: Text(L10().manufacturer),
           subtitle: Text(widget.supplierPart.manufacturerName),
           leading: FaIcon(FontAwesomeIcons.industry, color: COLOR_ACTION),
-          trailing: InvenTreeAPI().getImage(
-            widget.supplierPart.manufacturerImage,
-            width: 40,
-            height: 40,
-          ),
+          trailing: InvenTreeAPI().getThumbnail(widget.supplierPart.manufacturerImage),
           onTap: () async {
             showLoadingOverlay(context);
             var supplier = await InvenTreeCompany().get(widget.supplierPart.manufacturerId);
@@ -228,16 +217,6 @@ class _SupplierPartDisplayState extends RefreshableState<SupplierPartDetailWidge
     }
 
     return tiles;
-  }
-
-  @override
-  Widget getBody(BuildContext context) {
-    return ListView(
-        children: ListTile.divideTiles(
-          context: context,
-          tiles: detailTiles(context),
-        ).toList()
-    );
   }
 
 }